For Inthavong mesh ONLY
Will find boundary of olfactory region, by using regions IDed in the TECPLOT file. Should work even if points in the domain are modified (e.g. downsampled, upsampled, smoothed), as long as orientation/size of domain is the same, because each point will be mapped to its nearest face in the TECPLOT domain.